requirements.txt
openai
langchain
pinecone-client
python-dotenv
tiktoken
pypdf
wikipedia
